+++ build.gradle
@@ -1,4 +1,36 @@
+apply plugin: 'application'
 apply plugin: 'java'
+
+jar {
+    manifest {
+        attributes 'Main-Class': 'com.dalet.svnstats.Main'
+    }
+}
+
 repositories {
     mavenCentral()
+    mavenCentral artifactUrls: ["file://C:/users/mrabinovitch/.m2/repository/"]
 }
+
+dependencies {
+    compile(
+            [group: 'org.tmatesoft.svnkit', name: 'svnkit', version: '1.8.7'],
+            [group: 'log4j', name: 'log4j', version: '1.2.17'],
+            [group: 'org.hsqldb', name: 'hsqldb', version: '2.3.2'],
+            [files('C:/Program Files (x86)/IBM/Notes/jvm/lib/ext/Notes.jar')]
+    )
+    testCompile 'junit:junit:4.12'
+    runtime(
+            [group: 'org.tmatesoft.svnkit', name: 'svnkit', version: '1.8.7'],
+            [group: 'log4j', name: 'log4j', version: '1.2.17'],
+            [group: 'org.hsqldb', name: 'hsqldb', version: '2.3.2'],
+            [files('C:/Program Files (x86)/IBM/Notes/jvm/lib/ext/Notes.jar')]
+    )
+}
+
+mainClassName = 'com.dalet.svnstats.Main'
+
+task wrapper(type: Wrapper) {
+    gradleVersion = '2.4-rc-2'
+}
+
